Reorganize workflow schema
