+++ javascript/schema/src/index.ts
@@ -3,5 +3,4 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 export * from './tables.js'
-export * from './verbs.js'
-export * from './workflow.js'
+export * from './workflow/index.js'
