+++ javascript/schema/src/workflow/StepDefinition.ts
@@ -2,6 +2,7 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
+import type { BasicInput, DualInput,VariadicInput } from './bindings.js'
 import type {
 	AggregateArgs,
 	BinArgs,
@@ -36,73 +37,20 @@ import type {
 	WindowArgs,
 } from './verbs.js'
 
-export type PortBinding = string | NamedPortBinding
-export type OutputPortBinding = string | NamedOutputPortBinding
-
 /**
- * An explicit step input binding
+ * Common step properties
  */
-export interface NamedPortBinding {
+export interface StepJsonCommon {
 	/**
-	 * The id of the node to bind to
+	 * A unique identifier for this step
 	 */
-	node: string
-
-	/**
-	 * The named output of the node to bind with. If not defined, this will
-	 * be the default output "target"
-	 */
-	output?: string
-}
-
-/**
- * An explicit workflow output
- */
-export interface NamedOutputPortBinding extends NamedPortBinding {
-	/**
-	 * The output table name
-	 */
-	name: string
-}
-
-/**
- * The root wrangling workflow specification. (Used for generating JSON Schema)
- */
-export interface WorkflowJson {
-	/**
-	 * The schema url of the specification
-	 */
-	$schema?: string
-	/**
-	 * The name of the specification
-	 */
-	name?: string
-
-	/**
-	 * A user-friendly description of the specification
-	 */
-	description?: string
-
-	/**
-	 * The workflow steps
-	 */
-	steps?: StepJson[]
-
-	/**
-	 * A list of input names that are expected to be provided in addition to the workflow steps
-	 */
-	input?: string[]
-
-	/**
-	 * The output bindings
-	 */
-	output: Array<OutputPortBinding>
+	id?: string
 }
 
 /**
  * Specification for step items
  */
-export type StepJson = StepJsonCommon &
+export type StepDefinition = StepJsonCommon &
 	(
 		| ({ verb: Verb.Aggregate; args?: AggregateArgs } & BasicInput)
 		| ({ verb: Verb.Bin; args?: BinArgs } & BasicInput)
@@ -141,67 +89,3 @@ export type StepJson = StepJsonCommon &
 		| ({ verb: Verb.Unroll; args?: UnrollArgs } & BasicInput)
 		| ({ verb: Verb.Window; args?: WindowArgs } & BasicInput)
 	)
-
-/**
- * Common step properties
- */
-export interface StepJsonCommon {
-	/**
-	 * A unique identifier for this step
-	 */
-	id?: string
-}
-
-/**
- * Single-input, single-output step I/O
- */
-export interface BasicInput {
-	/**
-	 * Standard step input; single source with default name "source".
-	 *
-	 * If undefined, the default output of the previous step will be used (if available).
-	 * If no previous step is available, this will remain undefined
-	 */
-	input?: string | { source: PortBinding }
-}
-
-/**
- * Dual-input, single-output step I/O
- */
-export interface DualInput extends BasicInput {
-	/**
-	 * The inputs that must be bound; "source" & "other".
-	 */
-	input: {
-		/**
-		 * The primary input, which must be specified
-		 */
-		source: PortBinding
-
-		/**
-		 * The secondary input, which must be specified
-		 */
-		other: PortBinding
-	}
-}
-
-/**
- * Multi-input, single output step I/O
- */
-export interface VariadicInput extends BasicInput {
-	/**
-	 * The step inputs; a required "source" and optional, variadic "others". If this is a
-	 * string, it is used to bind the primary input.
-	 */
-	input: {
-		/**
-		 * The primary input
-		 */
-		source: PortBinding
-
-		/**
-		 * The variadic secondary inputs
-		 */
-		others?: PortBinding[]
-	}
-}
