+++ javascript/schema/src/workflow/index.ts
@@ -1,0 +1,8 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+export * from './bindings.js'
+export * from './StepDefinition.js'
+export * from './verbs.js'
+export * from './WorkflowDefinition.js'
