+++ javascript/schema/src/workflow/verbs.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import type { SortDirection, Value } from './tables.js'
+import type { SortDirection, Value } from '../tables.js'
 
 export enum Verb {
 	Aggregate = 'aggregate',
